# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetConnect API
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned after login or profile completion.
    SetupToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Setup token returned after OTP verification for onboarding.
  schemas:
...
    ServiceCatalog:
      type: object
      properties:
        id:
          type: string
          example: 64f1c3e2d2f2b70012a8b0aa
        name:
          type: string
          example: Full Grooming
        type:
          type: string
          enum: [vet, grooming, training, walking]
        price:
          type: number
          example: 45
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaxSettingResponse:
      type: object
      properties:
        percent:
          type: number
          example: 5
    AvailabilitySettingInput:
      type: object
      required: [startTime, endTime, slotMinutes]
      properties:
        startTime:
          type: string
          example: "09:00"
        endTime:
          type: string
          example: "17:00"
        slotMinutes:
          type: integer
          example: 30
    AvailabilitySettingResponse:
      type: object
      properties:
        startTime:
          type: string
          example: "09:00"
        endTime:
          type: string
          example: "17:00"
        slotMinutes:
          type: integer
          example: 30
...
    UpdateServicesInput:
      type: object
      required: [services]
      properties:
        services:
          type: array
          items:
            type: object
            required: [name, type, price]
            properties:
              name:
                type: string
                example: VET
              type:
                type: string
                enum: [vet, grooming, training, walking]
              price:
                type: number
                example: 250
              isActive:
                type: boolean
                example: true
...
paths:
  /admin/settings/services:
    get:
      tags: [AdminServiceSettings]
      summary: Get all service settings
      security:
        - AccessToken: []
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServiceCatalog"
        "401":
          description: Unauthorized
    patch:
      tags: [AdminServiceSettings]
      summary: Replace all service settings
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServicesInput"
      responses:
        "200":
          description: Services updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServiceCatalog"
        "400":
          description: Validation failed
    post:
      tags: [AdminServiceSettings]
      summary: Create service setting
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type, price]
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [vet, grooming, training, walking]
                price:
                  type: number
                isActive:
                  type: boolean
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/ServiceCatalog"
        "400":
          description: Validation failed
  /admin/settings/services/{id}:
    patch:
      tags: [AdminServiceSettings]
      summary: Update service setting
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                isActive:
                  type: boolean
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/ServiceCatalog"
        "404":
          description: Service not found
    delete:
      tags: [AdminServiceSettings]
      summary: Delete service setting
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "404":
          description: Service not found
  /admin/settings/tax:
    get:
      tags: [AdminServiceSettings]
      summary: Get active tax percentage
      security:
        - AccessToken: []
      responses:
        "200":
          description: Active tax
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/TaxSettingResponse"
    post:
      tags: [AdminServiceSettings]
      summary: Set active tax percentage
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [percent]
              properties:
                percent:
                  type: number
      responses:
        "201":
          description: Tax updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/TaxSettingResponse"
        "400":
          description: Validation failed
    patch:
      tags: [AdminServiceSettings]
      summary: Update active tax percentage
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [percent]
              properties:
                percent:
                  type: number
      responses:
        "200":
          description: Tax updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/TaxSettingResponse"
        "400":
          description: Validation failed
  /admin/settings/availability:
    get:
      tags: [AdminServiceSettings]
      summary: Get availability settings
      security:
        - AccessToken: []
      responses:
        "200":
          description: Availability settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/AvailabilitySettingResponse"
    patch:
      tags: [AdminServiceSettings]
      summary: Update availability settings
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilitySettingInput"
      responses:
        "200":
          description: Availability updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/AvailabilitySettingResponse"
        "400":
          description: Validation failed
  /services:
    get:
      tags: [UserServices]
      summary: Get available services and tax
      security:
        - AccessToken: []
      responses:
        "200":
          description: Available services
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      services:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceCatalog"
                      taxPercent:
                        type: number
        "401":
          description: Unauthorized
...